+++ crates/modor_ui/src/data.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
+pub enum StyleProperty<T> {
+    Fixed(T),
+    Manual,
+}