+++ examples/examples/ui.rs
@@ -1,0 +1,6 @@
+#![allow(missing_docs)]
+
+#[modor::modor_main]
+fn main() {
+    modor_examples::ui::main();
+}